# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LevelSummary {
  level: string;
  count: number;
}

export interface WordListItem {
  id: number;
  english: string;
  pos: string;
  level: string;
}

export interface Card {
  id: number;
  level: string;
  english: string;
  pos: string;
  arabic: string;
  sentenceEn: string;
  sentenceAr: string;
  audioWordUrl: string;
  audioSentenceUrl: string;
}

export type ListWordsParams = {
  level?: ListWordsLevel;
  search?: string;
};

export type ListWordsLevel =
  (typeof ListWordsLevel)[keyof typeof ListWordsLevel];

export const ListWordsLevel = {
  A1: "A1",
  A2: "A2",
  B1: "B1",
  B2: "B2",
  C1: "C1",
  ALL: "ALL",
} as const;